backend/requirements.txt
# FastAPI核心框架
fastapi==0.104.1
uvicorn[standard]==0.24.0
gunicorn==21.2.0

# 数据库相关
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
redis==5.0.1

# 文件处理
python-multipart==0.0.6
Pillow==10.1.0
ffmpeg-python==0.2.0
aiofiles==23.2.1

# 安全认证
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# 数据验证和序列化
pydantic==2.5.0
pydantic-settings==2.1.0

# 日志和工具
loguru==0.7.2
APScheduler==3.10.4

# 开发工具
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
